changed comment count to be a number instead of string

+++ models/app.model.js
@@ -27,7 +27,7 @@ exports.fetchAllArticles = () => {
         articles.created_at,
         articles.votes,
         articles.article_img_url, 
-        COUNT(comments.comment_id) AS comment_count
+        COUNT(comments.comment_id)::INT AS comment_count
         FROM articles
         LEFT JOIN comments
         ON articles.article_id = comments.article_id
